chore: rename get virtual asset list response dto to erase collection name in class
