+++ backend/core/src/main/java/com/scoinone/core/dto/response/virtualasset/GetVirtualAssetsResponseDto.java
